+++ packages/mui-material/src/styles/enhanceHighContrast.ts
