+++ packages/mui-material/src/styles/index.d.ts
@@ -8,7 +8,7 @@ export {
   Theme,
   CssThemeVariables,
 } from './createTheme';
-export { default as enhanceHighContrast, HighContrastTokens } from './createHighContrastTheme';
+export { default as enhanceHighContrast, HighContrastTokens } from './enhanceHighContrast';
 export { default as adaptV4Theme, DeprecatedThemeOptions } from './adaptV4Theme';
 export { Shadows } from './shadows';
 export { ZIndex } from './zIndex';
