+++ packages/mui-material/src/styles/index.js
@@ -25,7 +25,7 @@ export function experimental_sx() {
   );
 }
 export { default as createTheme } from './createTheme';
-export { default as enhanceHighContrast } from './createHighContrastTheme';
+export { default as enhanceHighContrast } from './enhanceHighContrast';
 export { default as unstable_createMuiStrictModeTheme } from './createMuiStrictModeTheme';
 export { default as createStyles } from './createStyles';
 export { getUnit as unstable_getUnit, toUnitless as unstable_toUnitless } from './cssUtils';
